SDK-1302: Nest modules to allow for tasks and filters
