--- spec/yoti/doc_scan/session/create/check_spec.rb
@@ -1,8 +1,0 @@
-describe 'Yoti::DocScan::Session::Create::Check' do
-  describe 'abstract class' do
-    it 'cannot be instantiated' do
-      expect { Yoti::DocScan::Session::Create::Check.new('', '') }
-        .to raise_error(TypeError, 'Yoti::DocScan::Session::Create::Check cannot be instantiated')
-    end
-  end
-end
